feat: add DFP quasi newton method
