+++ docs/api/minimise.md
@@ -32,9 +32,9 @@ In addition to the following, note that the [Optax](https://github.com/deepmind/
 
 ---
 
-??? abstract "`optimistix.AbstractBFGS`"
+??? abstract "`optimistix.AbstractQuasiNewton`"
 
-    ::: optimistix.AbstractBFGS
+    ::: optimistix.AbstractQuasiNewton
         options:
             members:
                 false
@@ -44,6 +44,31 @@ In addition to the following, note that the [Optax](https://github.com/deepmind/
         members:
             - __init__
 
+::: optimistix.DFP
+    selection:
+        members:
+            - __init__
+
+[`optimistix.AbstractQuasiNewton`][] supports several update methods for the Hessian approximation. optimistix currently provides two of these ([`optimistix.BFGSUpdate`][] and [`optimistix.DFPUpdate`][]), but you can also implement your own using the [`optimistix.AbstractQuasiNewtonUpdate`][] interface.
+
+??? abstract "`optimistix.AbstractQuasiNewtonUpdate`"
+
+    ::: optimistix.AbstractQuasiNewtonUpdate
+        selection:
+            members:
+                - __call__
+
+::: optimistix.BFGSUpdate
+    selection:
+        members:
+            - __init__
+
+::: optimistix.DFPUpdate
+    selection:
+        members:
+            - __init__
+
+
 ---
 
 ::: optimistix.OptaxMinimiser
@@ -51,7 +76,7 @@ In addition to the following, note that the [Optax](https://github.com/deepmind/
         members:
             - __init__
 
-`optim` in [`optimistix.OptaxMinimiser`][] is an instance of an Optax minimiser. For example, correct usage is `optimistix.OptaxMinimiser(optax.adam(...), ...)`, not `optimistix.OptaxMinimiser(optax.adam, ...)`. 
+`optim` in [`optimistix.OptaxMinimiser`][] is an instance of an Optax minimiser. For example, correct usage is `optimistix.OptaxMinimiser(optax.adam(...), ...)`, not `optimistix.OptaxMinimiser(optax.adam, ...)`.
 
 ---
