+++ optimistix/__init__.py
@@ -37,20 +37,24 @@ from ._search import (
 )
 from ._solution import RESULTS as RESULTS, Solution as Solution
 from ._solver import (
-    AbstractBFGS as AbstractBFGS,
     AbstractGaussNewton as AbstractGaussNewton,
     AbstractGradientDescent as AbstractGradientDescent,
+    AbstractQuasiNewton as AbstractQuasiNewton,
+    AbstractQuasiNewtonUpdate as AbstractQuasiNewtonUpdate,
     BacktrackingArmijo as BacktrackingArmijo,
     BestSoFarFixedPoint as BestSoFarFixedPoint,
     BestSoFarLeastSquares as BestSoFarLeastSquares,
     BestSoFarMinimiser as BestSoFarMinimiser,
     BestSoFarRootFinder as BestSoFarRootFinder,
     BFGS as BFGS,
+    BFGSUpdate as BFGSUpdate,
     Bisection as Bisection,
     Chord as Chord,
     ClassicalTrustRegion as ClassicalTrustRegion,
     dai_yuan as dai_yuan,
     DampedNewtonDescent as DampedNewtonDescent,
+    DFP as DFP,
+    DFPUpdate as DFPUpdate,
     Dogleg as Dogleg,
     DoglegDescent as DoglegDescent,
     FixedPointIteration as FixedPointIteration,
