+++ optimistix/_solver/quasi_newton.py
@@ -1,0 +1,594 @@
+import abc
+from collections.abc import Callable
+from typing import Any, cast, Generic, TypeVar, Union
+
+import equinox as eqx
+import jax
+import jax.numpy as jnp
+import jax.tree_util as jtu
+import lineax as lx
+from equinox import AbstractVar
+from equinox.internal import ω
+from jaxtyping import Array, Bool, Int, PyTree, Scalar
+
+from .._custom_types import Aux, DescentState, Fn, SearchState, Y
+from .._minimise import AbstractMinimiser
+from .._misc import (
+    cauchy_termination,
+    filter_cond,
+    lin_to_grad,
+    max_norm,
+    tree_dot,
+    tree_full_like,
+    verbose_print,
+)
+from .._search import (
+    AbstractDescent,
+    AbstractSearch,
+    FunctionInfo,
+)
+from .._solution import RESULTS
+from .backtracking import BacktrackingArmijo
+from .gauss_newton import NewtonDescent
+
+
+def _identity_pytree(pytree: PyTree[Array]) -> lx.PyTreeLinearOperator:
+    """Create an identity pytree `I` such that
+    `pytree = lx.PyTreeLinearOperator(I).mv(pytree)`
+
+    **Arguments**:
+
+    - `pytree`: A pytree such that the output of `_identity_pytree` is the identity
+        with respect to pytrees of the same shape as `pytree`.
+
+    **Returns**:
+
+    A `lx.PyTreeLinearOperator` with input and output shape the shape of `pytree`.
+    """
+    leaves, structure = jtu.tree_flatten(pytree)
+    eye_structure = structure.compose(structure)
+    eye_leaves = []
+    for i1, l1 in enumerate(leaves):
+        for i2, l2 in enumerate(leaves):
+            if i1 == i2:
+                eye_leaves.append(
+                    jnp.eye(jnp.size(l1)).reshape(jnp.shape(l1) + jnp.shape(l2))
+                )
+            else:
+                eye_leaves.append(jnp.zeros(jnp.shape(l1) + jnp.shape(l2)))
+
+    # This has a Lineax positive_semidefinite tag. This is okay because the BFGS update
+    # preserves positive-definiteness.
+    return lx.PyTreeLinearOperator(
+        jtu.tree_unflatten(eye_structure, eye_leaves),
+        jax.eval_shape(lambda: pytree),
+        lx.positive_semidefinite_tag,
+    )
+
+
+def _outer(tree1, tree2):
+    def leaf_fn(x):
+        return jtu.tree_map(lambda leaf: jnp.tensordot(x, leaf, axes=0), tree2)
+
+    return jtu.tree_map(leaf_fn, tree1)
+
+
+_Hessian = TypeVar(
+    "_Hessian", FunctionInfo.EvalGradHessian, FunctionInfo.EvalGradHessianInv
+)
+
+
+class AbstractQuasiNewtonUpdate(eqx.Module, strict=True):
+    """Abstract class for updating the approximate Hessian in a quasi-Newton method."""
+
+    use_inverse: AbstractVar[bool]
+
+    @abc.abstractmethod
+    def __call__(
+        self,
+        f_eval_info: FunctionInfo.EvalGrad,
+        f_info: Union[FunctionInfo.EvalGradHessian, FunctionInfo.EvalGradHessianInv],
+        y: PyTree,
+        y_eval: PyTree,
+    ) -> Union[FunctionInfo.EvalGradHessian, FunctionInfo.EvalGradHessianInv]:
+        ...
+
+
+class _AbstractBFGSDFPUpdate(AbstractQuasiNewtonUpdate, strict=True):
+    """Private intermediate class for BFGS/DFP updates."""
+
+    use_inverse: AbstractVar[bool]
+
+    def no_update(self, inner, grad_diff, y_diff, f_info):
+        if self.use_inverse:
+            return f_info.hessian_inv
+        else:
+            return f_info.hessian
+
+    @abc.abstractmethod
+    def update(
+        self,
+        inner: PyTree,
+        grad_diff: PyTree,
+        y_diff: PyTree,
+        f_info: Union[FunctionInfo.EvalGradHessian, FunctionInfo.EvalGradHessianInv],
+    ) -> lx.PyTreeLinearOperator:
+        ...
+
+    def __call__(
+        self,
+        f_eval_info: FunctionInfo.EvalGrad,
+        f_info: Union[FunctionInfo.EvalGradHessian, FunctionInfo.EvalGradHessianInv],
+        y: PyTree,
+        y_eval: PyTree,
+    ) -> Union[FunctionInfo.EvalGradHessian, FunctionInfo.EvalGradHessianInv]:
+        f_eval = f_eval_info.f
+        grad = f_eval_info.grad
+        y_diff = (y_eval**ω - y**ω).ω
+        grad_diff = (grad**ω - f_info.grad**ω).ω
+        inner = tree_dot(grad_diff, y_diff)
+
+        # In particular inner = 0 on the first step (as then state.grad=0), and so for
+        # this we jump straight to the line search.
+        # Likewise we get inner <= eps on convergence, and so again we make no update
+        # to avoid a division by zero.
+        inner_nonzero = inner > jnp.finfo(inner.dtype).eps
+
+        hessian = filter_cond(
+            inner_nonzero,
+            self.update,
+            self.no_update,
+            inner,
+            grad_diff,
+            y_diff,
+            f_info,
+        )
+        if self.use_inverse:
+            # in this case `hessian` is the new inverse hessian
+            return FunctionInfo.EvalGradHessianInv(f_eval, grad, hessian)
+        else:
+            return FunctionInfo.EvalGradHessian(f_eval, grad, hessian)
+
+
+class DFPUpdate(_AbstractBFGSDFPUpdate, strict=True):
+    """DFP (Davidon––Fletcher––Powell) approximate Hessian updates.
+
+    This is a variant of the BFGS update.
+
+    Ref: https://en.wikipedia.org/wiki/Davidon–Fletcher–Powell_formula.
+    """
+
+    use_inverse: bool
+
+    def update(self, inner, grad_diff, y_diff, f_info):
+        if self.use_inverse:
+            # Inverse Hessian
+            # pyright doesn't get it
+            f_info = cast(FunctionInfo.EvalGradHessianInv, f_info)
+            hessian_inv = f_info.hessian_inv
+            inv_mvp = hessian_inv.mv(grad_diff)
+            term1 = (_outer(y_diff, y_diff) ** ω / inner).ω
+            term2 = (_outer(inv_mvp, inv_mvp) ** ω / tree_dot(grad_diff, inv_mvp)).ω
+            new_hessian_inv = lx.PyTreeLinearOperator(
+                (hessian_inv.pytree**ω + term1**ω - term2**ω).ω,  # pyright: ignore
+                output_structure=jax.eval_shape(lambda: grad_diff),
+                tags=lx.positive_semidefinite_tag,
+            )
+            return new_hessian_inv
+        else:
+            # Hessian
+            # pyright doesn't get it
+            f_info = cast(FunctionInfo.EvalGradHessian, f_info)
+            hessian = f_info.hessian
+            mvp = hessian.mv(y_diff)
+            mvp_inner = tree_dot(y_diff, mvp)
+            diff_outer = _outer(grad_diff, grad_diff)
+            mvp_outer = _outer(grad_diff, mvp)
+            term1 = (((inner + mvp_inner) * (diff_outer**ω)) / (inner**2)).ω
+            term2 = ((_outer(mvp, grad_diff) ** ω + mvp_outer**ω) / inner).ω
+            new_hessian = lx.PyTreeLinearOperator(
+                (hessian.pytree**ω + term1**ω - term2**ω).ω,  # pyright: ignore
+                output_structure=jax.eval_shape(lambda: grad_diff),
+                tags=lx.positive_semidefinite_tag,
+            )
+            return new_hessian
+
+
+DFPUpdate.__init__.__doc__ = """**Arguments:**
+
+- `use_inverse`: The DFP algorithm involves computing matrix-vector products of the
+    form `B^{-1} g`, where `B` is an approximation to the Hessian of the function to be
+    minimised. This means we can either (a) store the approximate Hessian `B`, and do a
+    linear solve on every step, or (b) store the approximate Hessian inverse `B^{-1}`,
+    and do a matrix-vector product on every step. Option (a) is generally cheaper for
+    sparse Hessians (as the inverse may be dense). Option (b) is generally cheaper for
+    dense Hessians (as matrix-vector products are cheaper than linear solves). The
+    default is (b), denoted via `use_inverse=True`. Note that this is incompatible with
+    line search methods like [`optimistix.ClassicalTrustRegion`][], which use the
+    Hessian approximation `B` as part of their own computations.
+"""
+
+
+class BFGSUpdate(_AbstractBFGSDFPUpdate, strict=True):
+    """BFGS (Broyden--Fletcher--Goldfarb--Shanno) approximate Hessian updates.
+
+    Ref: https://en.wikipedia.org/wiki/Broyden–Fletcher–Goldfarb–Shanno_algorithm.
+    """
+
+    use_inverse: bool
+
+    def update(self, inner, grad_diff, y_diff, f_info):
+        if self.use_inverse:
+            # Inverse Hessian
+            # pyright doesn't get it
+            f_info = cast(FunctionInfo.EvalGradHessianInv, f_info)
+            hessian_inv = f_info.hessian_inv
+            # Use Woodbury identity for rank-1 update of approximate Hessian.
+            inv_mvp = hessian_inv.mv(grad_diff)
+            mvp_inner = tree_dot(grad_diff, inv_mvp)
+            diff_outer = _outer(y_diff, y_diff)
+            mvp_outer = _outer(y_diff, inv_mvp)
+            term1 = (((inner + mvp_inner) * (diff_outer**ω)) / (inner**2)).ω
+            term2 = ((_outer(inv_mvp, y_diff) ** ω + mvp_outer**ω) / inner).ω
+            new_hessian_inv = lx.PyTreeLinearOperator(
+                (hessian_inv.pytree**ω + term1**ω - term2**ω).ω,  # pyright: ignore
+                output_structure=jax.eval_shape(lambda: grad_diff),
+                tags=lx.positive_semidefinite_tag,
+            )
+            return new_hessian_inv
+        else:
+            # Hessian
+            # pyright doesn't get it
+            f_info = cast(FunctionInfo.EvalGradHessian, f_info)
+            hessian = f_info.hessian
+            # BFGS update to the operator directly
+            mvp = hessian.mv(y_diff)
+            term1 = (_outer(grad_diff, grad_diff) ** ω / inner).ω
+            term2 = (_outer(mvp, mvp) ** ω / tree_dot(y_diff, mvp)).ω
+            new_hessian = lx.PyTreeLinearOperator(
+                (hessian.pytree**ω + term1**ω - term2**ω).ω,  # pyright: ignore
+                output_structure=jax.eval_shape(lambda: grad_diff),
+                tags=lx.positive_semidefinite_tag,
+            )
+            return new_hessian
+
+
+BFGSUpdate.__init__.__doc__ = """**Arguments:**
+
+- `use_inverse`: The BFGS algorithm involves computing matrix-vector products of the
+    form `B^{-1} g`, where `B` is an approximation to the Hessian of the function to be
+    minimised. This means we can either (a) store the approximate Hessian `B`, and do a
+    linear solve on every step, or (b) store the approximate Hessian inverse `B^{-1}`,
+    and do a matrix-vector product on every step. Option (a) is generally cheaper for
+    sparse Hessians (as the inverse may be dense). Option (b) is generally cheaper for
+    dense Hessians (as matrix-vector products are cheaper than linear solves). The
+    default is (b), denoted via `use_inverse=True`. Note that this is incompatible with
+    line search methods like [`optimistix.ClassicalTrustRegion`][], which use the
+    Hessian approximation `B` as part of their own computations.
+"""
+
+
+class _QuasiNewtonState(
+    eqx.Module, Generic[Y, Aux, SearchState, DescentState, _Hessian], strict=True
+):
+    # Updated every search step
+    first_step: Bool[Array, ""]
+    y_eval: Y
+    search_state: SearchState
+    # Updated after each descent step
+    f_info: _Hessian
+    aux: Aux
+    descent_state: DescentState
+    # Used for termination
+    terminate: Bool[Array, ""]
+    result: RESULTS
+    # Used in compat.py
+    num_accepted_steps: Int[Array, ""]
+
+
+class AbstractQuasiNewton(
+    AbstractMinimiser[Y, Aux, _QuasiNewtonState], Generic[Y, Aux, _Hessian], strict=True
+):
+    """Abstract quasi-Newton minimisation algorithm.
+
+    This is a quasi-Newton optimisation algorithm, whose defining feature is the way
+    it progressively builds up a Hessian approximation using multiple steps of gradient
+    information. These updates to the Hessian are handled by the `hessian_update`
+    attribute, which should be a subclass of `AbstractQuasiNewtonUpdate`, such as
+    `BFGSUpdate` or `DFPUpdate`.
+
+    This abstract version may be subclassed to choose alternative descent and searches.
+
+    Supports the following `options`:
+
+    - `autodiff_mode`: whether to use forward- or reverse-mode autodifferentiation to
+        compute the gradient. Can be either `"fwd"` or `"bwd"`. Defaults to `"bwd"`,
+        which is usually more efficient. Changing this can be useful when the target
+        function does not support reverse-mode automatic differentiation.
+    """
+
+    rtol: AbstractVar[float]
+    atol: AbstractVar[float]
+    norm: AbstractVar[Callable[[PyTree], Scalar]]
+    descent: AbstractVar[AbstractDescent[Y, _Hessian, Any]]
+    search: AbstractVar[AbstractSearch[Y, _Hessian, FunctionInfo.Eval, Any]]
+    hessian_update: AbstractVar[AbstractQuasiNewtonUpdate]
+    verbose: AbstractVar[frozenset[str]]
+
+    def init(
+        self,
+        fn: Fn[Y, Scalar, Aux],
+        y: Y,
+        args: PyTree,
+        options: dict[str, Any],
+        f_struct: jax.ShapeDtypeStruct,
+        aux_struct: PyTree[jax.ShapeDtypeStruct],
+        tags: frozenset[object],
+    ) -> _QuasiNewtonState:
+        f = tree_full_like(f_struct, 0)
+        grad = tree_full_like(y, 0)
+        if self.hessian_update.use_inverse:
+            hessian_inv = _identity_pytree(y)
+            f_info = FunctionInfo.EvalGradHessianInv(f, grad, hessian_inv)
+        else:
+            hessian = _identity_pytree(y)
+            f_info = FunctionInfo.EvalGradHessian(f, grad, hessian)
+        f_info_struct = eqx.filter_eval_shape(lambda: f_info)
+        return _QuasiNewtonState(
+            first_step=jnp.array(True),
+            y_eval=y,
+            search_state=self.search.init(y, f_info_struct),
+            f_info=f_info,
+            aux=tree_full_like(aux_struct, 0),
+            descent_state=self.descent.init(y, f_info_struct),
+            terminate=jnp.array(False),
+            result=RESULTS.successful,
+            num_accepted_steps=jnp.array(0),
+        )
+
+    def step(
+        self,
+        fn: Fn[Y, Scalar, Aux],
+        y: Y,
+        args: PyTree,
+        options: dict[str, Any],
+        state: _QuasiNewtonState,
+        tags: frozenset[object],
+    ) -> tuple[Y, _QuasiNewtonState, Aux]:
+        autodiff_mode = options.get("autodiff_mode", "bwd")
+        f_eval, lin_fn, aux_eval = jax.linearize(
+            lambda _y: fn(_y, args), state.y_eval, has_aux=True
+        )
+        step_size, accept, search_result, search_state = self.search.step(
+            state.first_step,
+            y,
+            state.y_eval,
+            state.f_info,
+            FunctionInfo.Eval(f_eval),
+            state.search_state,
+        )
+
+        def accepted(descent_state):
+            grad = lin_to_grad(lin_fn, state.y_eval, autodiff_mode=autodiff_mode)
+
+            f_eval_info = self.hessian_update(
+                FunctionInfo.EvalGrad(f_eval, grad), state.f_info, y, state.y_eval
+            )
+
+            descent_state = self.descent.query(
+                state.y_eval,
+                f_eval_info,  # pyright: ignore
+                descent_state,
+            )
+            y_diff = (state.y_eval**ω - y**ω).ω
+            f_diff = (f_eval**ω - state.f_info.f**ω).ω
+            terminate = cauchy_termination(
+                self.rtol, self.atol, self.norm, state.y_eval, y_diff, f_eval, f_diff
+            )
+            terminate = jnp.where(
+                state.first_step, jnp.array(False), terminate
+            )  # Skip termination on first step
+            return state.y_eval, f_eval_info, aux_eval, descent_state, terminate
+
+        def rejected(descent_state):
+            return y, state.f_info, state.aux, descent_state, jnp.array(False)
+
+        y, f_info, aux, descent_state, terminate = filter_cond(
+            accept, accepted, rejected, state.descent_state
+        )
+
+        if len(self.verbose) > 0:
+            verbose_loss = "loss" in self.verbose
+            verbose_step_size = "step_size" in self.verbose
+            verbose_y = "y" in self.verbose
+            loss_eval = f_eval
+            loss = state.f_info.f
+            verbose_print(
+                (verbose_loss, "Loss on this step", loss_eval),
+                (verbose_loss, "Loss on the last accepted step", loss),
+                (verbose_step_size, "Step size", step_size),
+                (verbose_y, "y", state.y_eval),
+                (verbose_y, "y on the last accepted step", y),
+            )
+
+        y_descent, descent_result = self.descent.step(step_size, descent_state)
+        y_eval = (y**ω + y_descent**ω).ω
+        result = RESULTS.where(
+            search_result == RESULTS.successful, descent_result, search_result
+        )
+
+        state = _QuasiNewtonState(
+            first_step=jnp.array(False),
+            y_eval=y_eval,
+            search_state=search_state,
+            f_info=f_info,
+            aux=aux,
+            descent_state=descent_state,
+            terminate=terminate,
+            result=result,
+            num_accepted_steps=state.num_accepted_steps + jnp.where(accept, 1, 0),
+        )
+        return y, state, aux
+
+    def terminate(
+        self,
+        fn: Fn[Y, Scalar, Aux],
+        y: Y,
+        args: PyTree,
+        options: dict[str, Any],
+        state: _QuasiNewtonState,
+        tags: frozenset[object],
+    ) -> tuple[Bool[Array, ""], RESULTS]:
+        return state.terminate, state.result
+
+    def postprocess(
+        self,
+        fn: Fn[Y, Scalar, Aux],
+        y: Y,
+        aux: Aux,
+        args: PyTree,
+        options: dict[str, Any],
+        state: _QuasiNewtonState,
+        tags: frozenset[object],
+        result: RESULTS,
+    ) -> tuple[Y, Aux, dict[str, Any]]:
+        return y, aux, {}
+
+
+class BFGS(AbstractQuasiNewton[Y, Aux, _Hessian], strict=True):
+    """BFGS (Broyden--Fletcher--Goldfarb--Shanno) minimisation algorithm.
+
+    This is a quasi-Newton optimisation algorithm, whose defining feature is the way
+    it progressively builds up a Hessian approximation using multiple steps of gradient
+    information.
+
+    Supports the following `options`:
+
+    - `autodiff_mode`: whether to use forward- or reverse-mode autodifferentiation to
+        compute the gradient. Can be either `"fwd"` or `"bwd"`. Defaults to `"bwd"`,
+        which is usually more efficient. Changing this can be useful when the target
+        function does not support reverse-mode automatic differentiation.
+    """
+
+    rtol: float
+    atol: float
+    norm: Callable[[PyTree], Scalar]
+    descent: NewtonDescent
+    search: BacktrackingArmijo
+    hessian_update: AbstractQuasiNewtonUpdate
+    use_inverse: bool
+    verbose: frozenset[str]
+
+    def __init__(
+        self,
+        rtol: float,
+        atol: float,
+        norm: Callable[[PyTree], Scalar] = max_norm,
+        use_inverse: bool = True,
+        verbose: frozenset[str] = frozenset(),
+    ):
+        self.rtol = rtol
+        self.atol = atol
+        self.norm = norm
+        self.use_inverse = use_inverse
+        self.descent = NewtonDescent(linear_solver=lx.Cholesky())
+        # TODO(raderj): switch out `BacktrackingArmijo` with a better line search.
+        self.search = BacktrackingArmijo()
+        self.hessian_update = BFGSUpdate(use_inverse=use_inverse)
+        self.verbose = verbose
+
+
+BFGS.__init__.__doc__ = """**Arguments:**
+
+- `rtol`: Relative tolerance for terminating the solve.
+- `atol`: Absolute tolerance for terminating the solve.
+- `norm`: The norm used to determine the difference between two iterates in the
+    convergence criteria. Should be any function `PyTree -> Scalar`. Optimistix
+    includes three built-in norms: [`optimistix.max_norm`][],
+    [`optimistix.rms_norm`][], and [`optimistix.two_norm`][].
+- `use_inverse`: The BFGS algorithm involves computing matrix-vector products of the
+    form `B^{-1} g`, where `B` is an approximation to the Hessian of the function to be
+    minimised. This means we can either (a) store the approximate Hessian `B`, and do a
+    linear solve on every step, or (b) store the approximate Hessian inverse `B^{-1}`,
+    and do a matrix-vector product on every step. Option (a) is generally cheaper for
+    sparse Hessians (as the inverse may be dense). Option (b) is generally cheaper for
+    dense Hessians (as matrix-vector products are cheaper than linear solves). The
+    default is (b), denoted via `use_inverse=True`. Note that this is incompatible with
+    line search methods like [`optimistix.ClassicalTrustRegion`][], which use the
+    Hessian approximation `B` as part of their own computations.
+- `verbose`: Whether to print out extra information about how the solve is
+    proceeding. Should be a frozenset of strings, specifying what information to print.
+    Valid entries are `step_size`, `loss`, `y`. For example
+    `verbose=frozenset({"step_size", "loss"})`.
+"""
+
+
+class DFP(AbstractQuasiNewton[Y, Aux, _Hessian], strict=True):
+    """DFP (Davidon--Fletcher--Powell) minimisation algorithm.
+
+    This is a quasi-Newton optimisation algorithm, whose defining feature is the way
+    it progressively builds up a Hessian approximation using multiple steps of gradient
+    information.
+
+    [optimistix.BFGS][] is generally preferred, since it is more numerically stable on
+    most problems.
+
+    Supports the following `options`:
+
+    - `autodiff_mode`: whether to use forward- or reverse-mode autodifferentiation to
+        compute the gradient. Can be either `"fwd"` or `"bwd"`. Defaults to `"bwd"`,
+        which is usually more efficient. Changing this can be useful when the target
+        function does not support reverse-mode automatic differentiation.
+    """
+
+    rtol: float
+    atol: float
+    norm: Callable[[PyTree], Scalar]
+    descent: NewtonDescent
+    search: BacktrackingArmijo
+    hessian_update: AbstractQuasiNewtonUpdate
+    use_inverse: bool
+    verbose: frozenset[str]
+
+    def __init__(
+        self,
+        rtol: float,
+        atol: float,
+        norm: Callable[[PyTree], Scalar] = max_norm,
+        use_inverse: bool = True,
+        verbose: frozenset[str] = frozenset(),
+    ):
+        self.rtol = rtol
+        self.atol = atol
+        self.norm = norm
+        self.use_inverse = use_inverse
+        self.descent = NewtonDescent(linear_solver=lx.Cholesky())
+        # TODO(raderj): switch out `BacktrackingArmijo` with a better line search.
+        self.search = BacktrackingArmijo()
+        self.hessian_update = DFPUpdate(use_inverse=True)
+        self.verbose = verbose
+
+
+DFP.__init__.__doc__ = """**Arguments:**
+
+- `rtol`: Relative tolerance for terminating the solve.
+- `atol`: Absolute tolerance for terminating the solve.
+- `norm`: The norm used to determine the difference between two iterates in the
+    convergence criteria. Should be any function `PyTree -> Scalar`. Optimistix
+    includes three built-in norms: [`optimistix.max_norm`][],
+    [`optimistix.rms_norm`][], and [`optimistix.two_norm`][].
+- `use_inverse`: The DFP algorithm involves computing matrix-vector products of the
+    form `B^{-1} g`, where `B` is an approximation to the Hessian of the function to be
+    minimised. This means we can either (a) store the approximate Hessian `B`, and do a
+    linear solve on every step, or (b) store the approximate Hessian inverse `B^{-1}`,
+    and do a matrix-vector product on every step. Option (a) is generally cheaper for
+    sparse Hessians (as the inverse may be dense). Option (b) is generally cheaper for
+    dense Hessians (as matrix-vector products are cheaper than linear solves). The
+    default is (b), denoted via `use_inverse=True`. Note that this is incompatible with
+    line search methods like [`optimistix.ClassicalTrustRegion`][], which use the
+    Hessian approximation `B` as part of their own computations.
+- `verbose`: Whether to print out extra information about how the solve is
+    proceeding. Should be a frozenset of strings, specifying what information to print.
+    Valid entries are `step_size`, `loss`, `y`. For example
+    `verbose=frozenset({"step_size", "loss"})`.
+"""
