+++ tests/helpers.py
@@ -112,51 +112,63 @@ class DoglegMax(optx.AbstractGaussNewton[Y, Out, Aux]):
         self.verbose = frozenset()
 
 
-class BFGSDampedNewton(optx.AbstractBFGS):
+class BFGSDampedNewton(optx.AbstractQuasiNewton):
     """BFGS Hessian + direct Levenberg Marquardt update."""
 
     rtol: float
     atol: float
     norm: Callable = optx.max_norm
-    use_inverse: bool = False
     search: optx.AbstractSearch = optx.ClassicalTrustRegion()
     descent: optx.AbstractDescent = optx.DampedNewtonDescent()
+    hessian_update: optx.AbstractQuasiNewtonUpdate = optx.BFGSUpdate(use_inverse=False)
     verbose: frozenset[str] = frozenset()
 
 
-class BFGSIndirectDampedNewton(optx.AbstractBFGS):
+class BFGSIndirectDampedNewton(optx.AbstractQuasiNewton):
     """BFGS Hessian + indirect Levenberg Marquardt update."""
 
     rtol: float
     atol: float
     norm: Callable = optx.max_norm
-    use_inverse: bool = False
     search: optx.AbstractSearch = optx.ClassicalTrustRegion()
     descent: optx.AbstractDescent = optx.IndirectDampedNewtonDescent()
+    hessian_update: optx.AbstractQuasiNewtonUpdate = optx.BFGSUpdate(use_inverse=False)
     verbose: frozenset[str] = frozenset()
 
 
-class BFGSDogleg(optx.AbstractBFGS):
+class BFGSDogleg(optx.AbstractQuasiNewton):
     """BFGS Hessian + dogleg update."""
 
     rtol: float
     atol: float
     norm: Callable = optx.max_norm
-    use_inverse: bool = False
     search: optx.AbstractSearch = optx.ClassicalTrustRegion()
     descent: optx.AbstractDescent = optx.DoglegDescent(linear_solver=lx.SVD())
+    hessian_update: optx.AbstractQuasiNewtonUpdate = optx.BFGSUpdate(use_inverse=False)
     verbose: frozenset[str] = frozenset()
 
 
-class BFGSTrustRegion(optx.AbstractBFGS):
+class BFGSTrustRegion(optx.AbstractQuasiNewton):
     """Standard BFGS + classical trust region update."""
 
     rtol: float
     atol: float
     norm: Callable = optx.max_norm
-    use_inverse: bool = False
     search: optx.AbstractSearch = optx.LinearTrustRegion()
     descent: optx.AbstractDescent = optx.NewtonDescent()
+    hessian_update: optx.AbstractQuasiNewtonUpdate = optx.BFGSUpdate(use_inverse=True)
+    verbose: frozenset[str] = frozenset()
+
+
+class BFGSTrustRegionHessian(optx.AbstractQuasiNewton):
+    """Standard BFGS (uses hessian, not inverse!) + classical trust region update."""
+
+    rtol: float
+    atol: float
+    norm: Callable = optx.max_norm
+    search: optx.AbstractSearch = optx.LinearTrustRegion()
+    descent: optx.AbstractDescent = optx.NewtonDescent()
+    hessian_update: optx.AbstractQuasiNewtonUpdate = optx.BFGSUpdate(use_inverse=False)
     verbose: frozenset[str] = frozenset()
 
 
@@ -179,8 +191,10 @@ minimisers = (
     BFGSIndirectDampedNewton(rtol, atol),
     # Tighter tolerance needed to have bfgs_dogleg pass the JVP test.
     BFGSDogleg(1e-10, 1e-10),
-    BFGSTrustRegion(rtol, atol, use_inverse=False),
-    BFGSTrustRegion(rtol, atol, use_inverse=True),
+    BFGSTrustRegionHessian(rtol, atol),
+    BFGSTrustRegion(rtol, atol),
+    optx.DFP(rtol, atol, use_inverse=False),
+    optx.DFP(rtol, atol, use_inverse=True),
     optx.GradientDescent(1.5e-2, rtol, atol),
     # Tighter tolerance needed to have NonlinearCG pass the JVP test.
     optx.NonlinearCG(1e-10, 1e-10),
